+++ src/console/Application.php
@@ -31,6 +31,7 @@ class Application extends \yii\console\Application
         'health' => '\luya\console\commands\HealthController',
         'block' => '\luya\console\commands\BlockController',
         'storage' => '\luya\console\commands\StorageController',
+        'aw' => '\luya\console\commands\ActiveWindowController',
     ];
 
     /**